requirements.txt
alembic==0.8.10
ansible==2.7.6
apache-airflow==1.10.1
appnope==0.1.0
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.6
bleach==2.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.3
croniter==0.3.26
cryptography==2.5
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.9
docutils==0.14
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
hdfs3==0.3.0
html5lib==1.0.1
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.8.1
lockfile==0.12.2
lxml==4.3.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
mkl-fft==1.0.10
mkl-random==1.0.2
monotonic==1.5
netaddr==0.7.19
numpy==1.15.4
ordereddict==1.1
pandas==0.24.0
paramiko==2.4.2
parso==0.3.1
pendulum==1.4.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==4.4.2
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyarrow==0.11.1
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
pytzdata==2018.7
PyYAML==3.13
requests==2.21.0
setproctitle==1.1.10
sh==1.12.14
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.1.18
tabulate==0.8.2
tenacity==4.8.0
thrift==0.11.0
traitlets==4.3.2
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.4.0
